+++ demos/src/Examples/Menus/Vue/index.spec.js
@@ -1,7 +1,56 @@
-context('/src/Examples/BubbleMenu/Vue/', () => {
+context('/src/Examples/Menus/Vue/', () => {
   before(() => {
-    cy.visit('/src/Examples/BubbleMenu/Vue/')
+    cy.visit('/src/Examples/Menus/Vue/')
   })
 
-  // TODO: Write tests
+  beforeEach(() => {
+    cy.get('.ProseMirror').then(([{ editor }]) => {
+      editor.chain().focus().clearContent().run()
+    })
+  })
+
+  it('should show menu when the editor is empty', () => {
+    cy.get('#app')
+      .find('[data-tippy-root]')
+  })
+
+  it('should show menu when text is selected', () => {
+    cy.get('.ProseMirror')
+      .type('Test')
+      .type('{selectall}')
+
+    cy.get('#app')
+      .find('[data-tippy-root]')
+  })
+
+  const marks = [
+    {
+      button: 'Bold',
+      tag: 'strong',
+    },
+    {
+      button: 'Italic',
+      tag: 'em',
+    },
+    {
+      button: 'Strike',
+      tag: 's',
+    },
+  ]
+
+  marks.forEach(mark => {
+    it(`should apply ${mark.button} correctly`, () => {
+      cy.get('.ProseMirror')
+        .type('Test')
+        .type('{selectall}')
+
+      cy.get('#app div')
+        .find('[data-tippy-root]')
+        .contains(mark.button)
+        .click()
+
+      cy.get('.ProseMirror')
+        .find(`p ${mark.tag}`)
+    })
+  })
 })