+++ demos/src/Experiments/ExtensionStorage/React/CustomExtension.ts
@@ -1,0 +1,19 @@
+import { Extension } from '@tiptap/core'
+
+type CustomStorage = {
+  foo: number,
+}
+
+export const CustomExtension = Extension.create<{}, CustomStorage>({
+  name: 'custom',
+
+  addStorage() {
+    return {
+      foo: 123,
+    }
+  },
+
+  onUpdate() {
+    this.storage.foo += 1
+  },
+})