+++ packages/core/src/Node.ts
@@ -8,7 +8,10 @@ import { Plugin, Transaction } from 'prosemirror-state'
 import { InputRule } from './InputRule'
 import { PasteRule } from './PasteRule'
 import mergeDeep from './utilities/mergeDeep'
+import callOrReturn from './utilities/callOrReturn'
+import getExtensionField from './helpers/getExtensionField'
 import {
+  AnyConfig,
   Extensions,
   Attributes,
   NodeViewRenderer,
@@ -21,7 +24,7 @@ import { NodeConfig } from '.'
 import { Editor } from './Editor'
 
 declare module '@tiptap/core' {
-  interface NodeConfig<Options = any> {
+  interface NodeConfig<Options = any, Storage = any> {
     [key: string]: any;
 
     /**
@@ -40,12 +43,22 @@ declare module '@tiptap/core' {
     defaultOptions?: Options,
 
     /**
+     * Default Storage
+     */
+    addStorage?: (this: {
+      name: string,
+      options: Options,
+      parent: ParentConfig<NodeConfig<Options, Storage>>['addGlobalAttributes'],
+    }) => Storage,
+
+    /**
      * Global attributes
      */
     addGlobalAttributes?: (this: {
       name: string,
       options: Options,
-      parent: ParentConfig<NodeConfig<Options>>['addGlobalAttributes'],
+      storage: Storage,
+      parent: ParentConfig<NodeConfig<Options, Storage>>['addGlobalAttributes'],
     }) => GlobalAttributes | {},
 
     /**
@@ -54,9 +67,10 @@ declare module '@tiptap/core' {
     addCommands?: (this: {
       name: string,
       options: Options,
+      storage: Storage,
       editor: Editor,
       type: NodeType,
-      parent: ParentConfig<NodeConfig<Options>>['addCommands'],
+      parent: ParentConfig<NodeConfig<Options, Storage>>['addCommands'],
     }) => Partial<RawCommands>,
 
     /**
@@ -65,9 +79,10 @@ declare module '@tiptap/core' {
     addKeyboardShortcuts?: (this: {
       name: string,
       options: Options,
+      storage: Storage,
       editor: Editor,
       type: NodeType,
-      parent: ParentConfig<NodeConfig<Options>>['addKeyboardShortcuts'],
+      parent: ParentConfig<NodeConfig<Options, Storage>>['addKeyboardShortcuts'],
     }) => {
       [key: string]: KeyboardShortcutCommand,
     },
@@ -78,9 +93,10 @@ declare module '@tiptap/core' {
     addInputRules?: (this: {
       name: string,
       options: Options,
+      storage: Storage,
       editor: Editor,
       type: NodeType,
-      parent: ParentConfig<NodeConfig<Options>>['addInputRules'],
+      parent: ParentConfig<NodeConfig<Options, Storage>>['addInputRules'],
     }) => InputRule[],
 
     /**
@@ -89,9 +105,10 @@ declare module '@tiptap/core' {
     addPasteRules?: (this: {
       name: string,
       options: Options,
+      storage: Storage,
       editor: Editor,
       type: NodeType,
-      parent: ParentConfig<NodeConfig<Options>>['addPasteRules'],
+      parent: ParentConfig<NodeConfig<Options, Storage>>['addPasteRules'],
     }) => PasteRule[],
 
     /**
@@ -100,9 +117,10 @@ declare module '@tiptap/core' {
     addProseMirrorPlugins?: (this: {
       name: string,
       options: Options,
+      storage: Storage,
       editor: Editor,
       type: NodeType,
-      parent: ParentConfig<NodeConfig<Options>>['addProseMirrorPlugins'],
+      parent: ParentConfig<NodeConfig<Options, Storage>>['addProseMirrorPlugins'],
     }) => Plugin[],
 
     /**
@@ -111,7 +129,8 @@ declare module '@tiptap/core' {
     addExtensions?: (this: {
       name: string,
       options: Options,
-      parent: ParentConfig<NodeConfig<Options>>['addExtensions'],
+      storage: Storage,
+      parent: ParentConfig<NodeConfig<Options, Storage>>['addExtensions'],
     }) => Extensions,
 
     /**
@@ -121,7 +140,8 @@ declare module '@tiptap/core' {
       this: {
         name: string,
         options: Options,
-        parent: ParentConfig<NodeConfig<Options>>['extendNodeSchema'],
+        storage: Storage,
+        parent: ParentConfig<NodeConfig<Options, Storage>>['extendNodeSchema'],
       },
       extension: Node,
     ) => Record<string, any>) | null,
@@ -133,7 +153,8 @@ declare module '@tiptap/core' {
       this: {
         name: string,
         options: Options,
-        parent: ParentConfig<NodeConfig<Options>>['extendMarkSchema'],
+        storage: Storage,
+        parent: ParentConfig<NodeConfig<Options, Storage>>['extendMarkSchema'],
       },
       extension: Node,
     ) => Record<string, any>) | null,
@@ -144,9 +165,10 @@ declare module '@tiptap/core' {
     onBeforeCreate?: ((this: {
       name: string,
       options: Options,
+      storage: Storage,
       editor: Editor,
       type: NodeType,
-      parent: ParentConfig<NodeConfig<Options>>['onBeforeCreate'],
+      parent: ParentConfig<NodeConfig<Options, Storage>>['onBeforeCreate'],
     }) => void) | null,
 
     /**
@@ -155,9 +177,10 @@ declare module '@tiptap/core' {
     onCreate?: ((this: {
       name: string,
       options: Options,
+      storage: Storage,
       editor: Editor,
       type: NodeType,
-      parent: ParentConfig<NodeConfig<Options>>['onCreate'],
+      parent: ParentConfig<NodeConfig<Options, Storage>>['onCreate'],
     }) => void) | null,
 
     /**
@@ -166,9 +189,10 @@ declare module '@tiptap/core' {
     onUpdate?: ((this: {
       name: string,
       options: Options,
+      storage: Storage,
       editor: Editor,
       type: NodeType,
-      parent: ParentConfig<NodeConfig<Options>>['onUpdate'],
+      parent: ParentConfig<NodeConfig<Options, Storage>>['onUpdate'],
     }) => void) | null,
 
     /**
@@ -177,9 +201,10 @@ declare module '@tiptap/core' {
     onSelectionUpdate?: ((this: {
       name: string,
       options: Options,
+      storage: Storage,
       editor: Editor,
       type: NodeType,
-      parent: ParentConfig<NodeConfig<Options>>['onSelectionUpdate'],
+      parent: ParentConfig<NodeConfig<Options, Storage>>['onSelectionUpdate'],
     }) => void) | null,
 
     /**
@@ -189,9 +214,10 @@ declare module '@tiptap/core' {
       this: {
         name: string,
         options: Options,
+        storage: Storage,
         editor: Editor,
         type: NodeType,
-        parent: ParentConfig<NodeConfig<Options>>['onTransaction'],
+        parent: ParentConfig<NodeConfig<Options, Storage>>['onTransaction'],
       },
       props: {
         transaction: Transaction,
@@ -205,9 +231,10 @@ declare module '@tiptap/core' {
       this: {
         name: string,
         options: Options,
+        storage: Storage,
         editor: Editor,
         type: NodeType,
-        parent: ParentConfig<NodeConfig<Options>>['onFocus'],
+        parent: ParentConfig<NodeConfig<Options, Storage>>['onFocus'],
       },
       props: {
         event: FocusEvent,
@@ -221,9 +248,10 @@ declare module '@tiptap/core' {
       this: {
         name: string,
         options: Options,
+        storage: Storage,
         editor: Editor,
         type: NodeType,
-        parent: ParentConfig<NodeConfig<Options>>['onBlur'],
+        parent: ParentConfig<NodeConfig<Options, Storage>>['onBlur'],
       },
       props: {
         event: FocusEvent,
@@ -236,9 +264,10 @@ declare module '@tiptap/core' {
     onDestroy?: ((this: {
       name: string,
       options: Options,
+      storage: Storage,
       editor: Editor,
       type: NodeType,
-      parent: ParentConfig<NodeConfig<Options>>['onDestroy'],
+      parent: ParentConfig<NodeConfig<Options, Storage>>['onDestroy'],
     }) => void) | null,
 
     /**
@@ -247,9 +276,10 @@ declare module '@tiptap/core' {
     addNodeView?: ((this: {
       name: string,
       options: Options,
+      storage: Storage,
       editor: Editor,
       type: NodeType,
-      parent: ParentConfig<NodeConfig<Options>>['addNodeView'],
+      parent: ParentConfig<NodeConfig<Options, Storage>>['addNodeView'],
     }) => NodeViewRenderer) | null,
 
     /**
@@ -263,7 +293,8 @@ declare module '@tiptap/core' {
     content?: NodeSpec['content'] | ((this: {
       name: string,
       options: Options,
-      parent: ParentConfig<NodeConfig<Options>>['content'],
+      storage: Storage,
+      parent: ParentConfig<NodeConfig<Options, Storage>>['content'],
     }) => NodeSpec['content']),
 
     /**
@@ -272,7 +303,8 @@ declare module '@tiptap/core' {
     marks?: NodeSpec['marks'] | ((this: {
       name: string,
       options: Options,
-      parent: ParentConfig<NodeConfig<Options>>['marks'],
+      storage: Storage,
+      parent: ParentConfig<NodeConfig<Options, Storage>>['marks'],
     }) => NodeSpec['marks']),
 
     /**
@@ -281,7 +313,8 @@ declare module '@tiptap/core' {
     group?: NodeSpec['group'] | ((this: {
       name: string,
       options: Options,
-      parent: ParentConfig<NodeConfig<Options>>['group'],
+      storage: Storage,
+      parent: ParentConfig<NodeConfig<Options, Storage>>['group'],
     }) => NodeSpec['group']),
 
     /**
@@ -290,7 +323,8 @@ declare module '@tiptap/core' {
     inline?: NodeSpec['inline'] | ((this: {
       name: string,
       options: Options,
-      parent: ParentConfig<NodeConfig<Options>>['inline'],
+      storage: Storage,
+      parent: ParentConfig<NodeConfig<Options, Storage>>['inline'],
     }) => NodeSpec['inline']),
 
     /**
@@ -299,7 +333,8 @@ declare module '@tiptap/core' {
     atom?: NodeSpec['atom'] | ((this: {
       name: string,
       options: Options,
-      parent: ParentConfig<NodeConfig<Options>>['atom'],
+      storage: Storage,
+      parent: ParentConfig<NodeConfig<Options, Storage>>['atom'],
     }) => NodeSpec['atom']),
 
     /**
@@ -308,7 +343,8 @@ declare module '@tiptap/core' {
     selectable?: NodeSpec['selectable'] | ((this: {
       name: string,
       options: Options,
-      parent: ParentConfig<NodeConfig<Options>>['selectable'],
+      storage: Storage,
+      parent: ParentConfig<NodeConfig<Options, Storage>>['selectable'],
     }) => NodeSpec['selectable']),
 
     /**
@@ -317,7 +353,8 @@ declare module '@tiptap/core' {
     draggable?: NodeSpec['draggable'] | ((this: {
       name: string,
       options: Options,
-      parent: ParentConfig<NodeConfig<Options>>['draggable'],
+      storage: Storage,
+      parent: ParentConfig<NodeConfig<Options, Storage>>['draggable'],
     }) => NodeSpec['draggable']),
 
     /**
@@ -326,7 +363,8 @@ declare module '@tiptap/core' {
     code?: NodeSpec['code'] | ((this: {
       name: string,
       options: Options,
-      parent: ParentConfig<NodeConfig<Options>>['code'],
+      storage: Storage,
+      parent: ParentConfig<NodeConfig<Options, Storage>>['code'],
     }) => NodeSpec['code']),
 
     /**
@@ -335,7 +373,8 @@ declare module '@tiptap/core' {
     defining?: NodeSpec['defining'] | ((this: {
       name: string,
       options: Options,
-      parent: ParentConfig<NodeConfig<Options>>['defining'],
+      storage: Storage,
+      parent: ParentConfig<NodeConfig<Options, Storage>>['defining'],
     }) => NodeSpec['defining']),
 
     /**
@@ -344,7 +383,8 @@ declare module '@tiptap/core' {
     isolating?: NodeSpec['isolating'] | ((this: {
       name: string,
       options: Options,
-      parent: ParentConfig<NodeConfig<Options>>['isolating'],
+      storage: Storage,
+      parent: ParentConfig<NodeConfig<Options, Storage>>['isolating'],
     }) => NodeSpec['isolating']),
 
     /**
@@ -354,7 +394,8 @@ declare module '@tiptap/core' {
       this: {
         name: string,
         options: Options,
-        parent: ParentConfig<NodeConfig<Options>>['parseHTML'],
+        storage: Storage,
+        parent: ParentConfig<NodeConfig<Options, Storage>>['parseHTML'],
       },
     ) => NodeSpec['parseDOM'],
 
@@ -365,7 +406,8 @@ declare module '@tiptap/core' {
       this: {
         name: string,
         options: Options,
-        parent: ParentConfig<NodeConfig<Options>>['renderHTML'],
+        storage: Storage,
+        parent: ParentConfig<NodeConfig<Options, Storage>>['renderHTML'],
       },
       props: {
         node: ProseMirrorNode,
@@ -380,7 +422,8 @@ declare module '@tiptap/core' {
       this: {
         name: string,
         options: Options,
-        parent: ParentConfig<NodeConfig<Options>>['renderText'],
+        storage: Storage,
+        parent: ParentConfig<NodeConfig<Options, Storage>>['renderText'],
       },
       props: {
         node: ProseMirrorNode,
@@ -397,13 +440,14 @@ declare module '@tiptap/core' {
       this: {
         name: string,
         options: Options,
-        parent: ParentConfig<NodeConfig<Options>>['addAttributes'],
+        storage: Storage,
+        parent: ParentConfig<NodeConfig<Options, Storage>>['addAttributes'],
       },
     ) => Attributes | {},
   }
 }
 
-export class Node<Options = any> {
+export class Node<Options = any, Storage = any> {
   type = 'node'
 
   name = 'node'
@@ -414,12 +458,14 @@ export class Node<Options = any> {
 
   options: Options
 
+  storage: Storage
+
   config: NodeConfig = {
     name: this.name,
     defaultOptions: {},
   }
 
-  constructor(config: Partial<NodeConfig<Options>> = {}) {
+  constructor(config: Partial<NodeConfig<Options, Storage>> = {}) {
     this.config = {
       ...this.config,
       ...config,
@@ -427,10 +473,18 @@ export class Node<Options = any> {
 
     this.name = this.config.name
     this.options = this.config.defaultOptions
+    this.storage = callOrReturn(getExtensionField<AnyConfig['addStorage']>(
+      this,
+      'addStorage',
+      {
+        name: this.name,
+        options: this.options,
+      },
+    ))
   }
 
-  static create<O>(config: Partial<NodeConfig<O>> = {}) {
-    return new Node<O>(config)
+  static create<O = any, S = any>(config: Partial<NodeConfig<O, S>> = {}) {
+    return new Node<O, S>(config)
   }
 
   configure(options: Partial<Options> = {}) {
@@ -440,11 +494,20 @@ export class Node<Options = any> {
 
     extension.options = mergeDeep(this.options, options) as Options
 
+    extension.storage = callOrReturn(getExtensionField<AnyConfig['addStorage']>(
+      extension,
+      'addStorage',
+      {
+        name: extension.name,
+        options: extension.options,
+      },
+    ))
+
     return extension
   }
 
-  extend<ExtendedOptions = Options>(extendedConfig: Partial<NodeConfig<ExtendedOptions>> = {}) {
-    const extension = new Node<ExtendedOptions>(extendedConfig)
+  extend<ExtendedOptions = Options, ExtendedStorage = Storage>(extendedConfig: Partial<NodeConfig<ExtendedOptions, ExtendedStorage>> = {}) {
+    const extension = new Node<ExtendedOptions, ExtendedStorage>(extendedConfig)
 
     extension.parent = this
 
@@ -457,6 +520,15 @@ export class Node<Options = any> {
     extension.options = extendedConfig.defaultOptions
       ? extendedConfig.defaultOptions
       : extension.parent.options
+
+    extension.storage = callOrReturn(getExtensionField<AnyConfig['addStorage']>(
+      extension,
+      'addStorage',
+      {
+        name: extension.name,
+        options: extension.options,
+      },
+    ))
 
     return extension
   }