+++ packages/core/src/utilities/findDuplicates.ts
@@ -1,0 +1,5 @@
+export default function findDuplicates(items: any[]): any[] {
+  const filtered = items.filter((el, index) => items.indexOf(el) !== index)
+
+  return [...new Set(filtered)]
+}