+++ packages/extension-table/src/table.ts
@@ -66,13 +66,14 @@ declare module '@tiptap/core' {
     }
   }
 
-  interface NodeConfig<Options> {
+  interface NodeConfig<Options, Storage> {
     /**
      * Table Role
      */
     tableRole?: string | ((this: {
       name: string,
       options: Options,
+      storage: Storage,
       parent: ParentConfig<NodeConfig<Options>>['tableRole'],
     }) => string),
   }
@@ -245,6 +246,7 @@ export const Table = Node.create<TableOptions>({
     const context = {
       name: extension.name,
       options: extension.options,
+      storage: extension.storage,
     }
 
     return {